% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.merge.R
\name{multi.merge}
\alias{multi.merge}
\title{warning as they appear, otherwise, R throw all the warnings at the end
This package merges multiple dfs into one with a small number of arguments.
Apart from the arguments needed in this package, the default values from
base::merge are used.
The errors/warnings reported from this package are limited, as I have dependent
upon the errors/warnings from the base::merge.}
\usage{
multi.merge(DFlist, byCol, type = "union")
}
\arguments{
\item{byCol}{A string of the name of the column use to merge the data frames, currently it supports single column to be specified}

\item{type}{The type of merge as string, either one of intersect, left, right, union.}

\item{infile}{The list of data frames to be merged}
}
\value{
A data frame of merged data frames in the list.
}
\description{
"byCol" argument can accept the column number used for merging, however, it is
not an intended functionality.
}
\details{
The merge "type" argument can match the partial merge-type, which is not recommended,
although, there are only four types available, the partial match functionality is
solely for typos.
}
